# scripts/maintenance/openapi_baseline.json
{
  "components": {
    "schemas": {
...
      "BookMetaEntryOut": {
        "description": "meta2.db の書籍メタデータ 1 件。",
        "properties": {
          "asin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asin"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "hidden": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hidden"
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "last_viewed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Viewed At"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "read_state": {
            "anyOf": [
              {
                "enum": [
                  "unread",
                  "reading",
                  "done"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read State"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "series_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Index"
          },
          "series_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Title"
          },
          "view_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "authors"
        ],
        "title": "BookMetaEntryOut",
        "type": "object"
      },
...
      "BuildEnqueueResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "queued_position": {
            "title": "Queued Position",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "queued_position"
        ],
        "title": "BuildEnqueueResponse",
        "type": "object"
      },
...
      "EnqueueRequest": {
        "properties": {
          "all_books": {
            "default": false,
            "title": "All Books",
            "type": "boolean"
          },
          "book_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Name"
          },
          "mode": {
            "default": "full_build",
            "enum": [
              "full_build",
              "generate_contexts",
              "generate_relations"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "title": "EnqueueRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetaUpdateResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "updated_count"
        ],
        "title": "MetaUpdateResponse",
        "type": "object"
      },
...
      "RecordViewRequest": {
        "description": "閲覧記録リクエスト。",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "source": {
            "default": "doujin",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RecordViewRequest",
        "type": "object"
      },
      "RecordViewResponse": {
        "properties": {
          "incremented": {
            "title": "Incremented",
            "type": "boolean"
          },
          "last_viewed_at": {
            "title": "Last Viewed At",
            "type": "number"
          },
          "read_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read State"
          },
          "view_count": {
            "title": "View Count",
            "type": "integer"
          }
        },
        "required": [
          "view_count",
          "last_viewed_at",
          "incremented"
        ],
        "title": "RecordViewResponse",
        "type": "object"
      },
...
      "UpdateMetaRequest": {
        "description": "単一書籍または複数書籍へのメタデータ更新リクエスト。\n\n`authors` / `hidden` / `genre` / `read_state` は省略可。省略されたフィールドは変更しない。\nすべて省略するとエラー（更新する内容が無い）。",
        "properties": {
          "authors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authors"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "hidden": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hidden"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "title": "Names",
            "type": "array"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "read_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read State"
          },
          "source": {
            "default": "doujin",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "names"
        ],
        "title": "UpdateMetaRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "WebP / ZIP 画像を PDF 化してブラウザ閲覧するマルチソース（doujin / comic / novel）ライブラリのバックエンド API。Kindle キャプチャ連携、yomitoku による OCR 検索可能 PDF 生成、novel 向けの bge-m3 + Qwen によるフルテキスト検索・RAG 質問応答を含む。",
    "title": "Pic2PDF Viewer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meta": {
      "get": {
        "description": "指定ソースの meta.json 全体を返す。\nレスポンス: { \"key\": { \"authors\": [...] }, ... }",
        "operationId": "get_meta_api_meta_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "doujin",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/BookMetaEntryOut"
                  },
                  "title": "Response Get Meta Api Meta Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Meta",
        "tags": [
          "meta"
        ]
      },
      "patch": {
        "description": "1冊または複数冊のメタデータ（作者名 / 非表示フラグ / ジャンル / 読書状態）を上書き保存する。\n\n指定したフィールドのみ書き換え、省略されたフィールドは保持する。\nすべての list フィールドが空になった場合はエントリ自体を削除する。",
        "operationId": "update_meta_api_meta_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMetaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Meta",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/meta/export": {
      "get": {
        "description": "指定ソースの meta.json 全体を JSON ファイルとしてダウンロードする。\nバックアップ・環境移行用。副作用なし（読み取り専用）。",
        "operationId": "export_meta_api_meta_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "doujin",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Meta",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/api/meta/view": {
      "post": {
        "description": "書籍の閲覧を記録する。\n\n- `last_viewed_at` は呼び出し毎に常に現在時刻で更新（最近見た順ソート用）。\n- `view_count` は前回 last_viewed_at から VIEW_COUNT_DEBOUNCE_SEC 以上経過した場合のみ +1。\n  短時間で同じ書籍を何度も開いてもカウントが膨らまないようにする。",
        "operationId": "record_view_api_meta_view_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordViewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record View",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/api/novel/build/enqueue": {
      "post": {
        "description": "Full Build / コンテキスト生成ジョブをキューに登録する（API §8.1）。",
        "operationId": "post_enqueue_api_novel_build_enqueue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildEnqueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Enqueue",
        "tags": [
          "novel_build"
        ]
      }
    },
    "/api/novel/build/jobs/{job_id}": {
      "delete": {
        "description": "待機中 Full Build ジョブをキャンセルする（API §8.3）。",
        "operationId": "delete_job_api_novel_build_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "novel_build"
        ]
      }
    },
...
  }
}